Move Azure Methods to Own Class
Extract the Azure SQL Server configuration methods from the standard
extensions class into their own space so that it is simpler to read and
maintain.
