+++ src/dbup-sqlserver/SqlServerExtensions.cs
@@ -43,61 +43,6 @@ public static class SqlServerExtensions
         return SqlDatabase(new SqlConnectionManager(connectionString), schema);
     }
 
-#if SUPPORTS_AZURE_AD
-    /// <summary>
-    /// Creates an upgrader for SQL Server databases.
-    /// </summary>
-    /// <param name="supported">Fluent helper type.</param>
-    /// <param name="connectionString">The connection string.</param>
-    /// <param name="schema">The SQL schema name to use. Defaults to 'dbo' if <see langword="null" />.</param>
-    /// <param name="useAzureSqlIntegratedSecurity">Whether to use Azure SQL Integrated Security</param>
-    /// <returns>
-    /// A builder for a database upgrader designed for SQL Server databases.
-    /// </returns>
-    public static UpgradeEngineBuilder SqlDatabase(this SupportedDatabases supported, string connectionString, string schema, bool useAzureSqlIntegratedSecurity)
-    {
-        if (useAzureSqlIntegratedSecurity)
-        {
-            return AzureSqlDatabaseWithIntegratedSecurity(supported, connectionString, schema);
-        }
-
-        return SqlDatabase(new SqlConnectionManager(connectionString), schema);
-    }
-
-    /// <summary>Creates an upgrader for Azure SQL Databases using Azure AD Integrated Security.</summary>
-    /// <param name="supported">Fluent helper type.</param>
-    /// <param name="connectionString">The connection string.</param>
-    /// <param name="schema">The SQL schema name to use. Defaults to 'dbo' if <see langword="null" />.</param>
-    /// <returns>A builder for a database upgrader designed for Azure SQL Server databases.</returns>
-    public static UpgradeEngineBuilder AzureSqlDatabaseWithIntegratedSecurity(this SupportedDatabases supported, string connectionString, string schema)
-    {
-        return SqlDatabase(new AzureSqlConnectionManager(connectionString), schema);
-    }
-
-    /// <summary>Creates an upgrader for Azure SQL Databases using Azure AD Integrated Security.</summary>
-    /// <param name="supported">Fluent helper type.</param>
-    /// <param name="connectionString">The connection string.</param>
-    /// <param name="schema">The SQL schema name to use. Defaults to 'dbo' if <see langword="null" />.</param>
-    /// <returns>A builder for a database upgrader designed for Azure SQL Server databases.</returns>
-    public static UpgradeEngineBuilder AzureSqlDatabaseWithIntegratedSecurity(this SupportedDatabases supported, string connectionString, string schema, string resource)
-    {
-        return AzureSqlDatabaseWithIntegratedSecurity(supported, connectionString, schema, resource, null);
-    }
-
-    /// <summary>Creates an upgrader for Azure SQL Databases using Azure AD Integrated Security.</summary>
-    /// <param name="supported">Fluent helper type.</param>
-    /// <param name="connectionString">The connection string.</param>
-    /// <param name="schema">The SQL schema name to use. Defaults to 'dbo' if <see langword="null" />.</param>
-    /// <param name="azureAdInstance">Specify a value for clouds other than the Public Cloud.</param>
-    /// <param name="resource">Resource to access. e.g. https://management.azure.com/.</param>
-    /// <param name="tenantId">If not specified, default tenant is used. Managed Service Identity REST protocols do not accept tenantId, so this can only be used with certificate and client secret based authentication.</param>
-    /// <returns>A builder for a database upgrader designed for Azure SQL Server databases.</returns>
-    public static UpgradeEngineBuilder AzureSqlDatabaseWithIntegratedSecurity(this SupportedDatabases supported, string connectionString, string schema, string resource, string tenantId, string azureAdInstance = "https://login.microsoftonline.com/")
-    {
-        return SqlDatabase(new AzureSqlConnectionManager(connectionString, azureAdInstance, resource, tenantId), schema);
-    }
-#endif
-
     /// <summary>
     /// Creates an upgrader for SQL Server databases.
     /// </summary>
